TZ, BJ, SL, RW added
Along with the new scripts a number of functions were added to the tools script to aggregate across administrative levels (admin 2 to admin 1) after checking that both harvested area and production are available for all observations and that at least 50% of the production in a given admin 1 is represented by the admin 2 levels available. This ensures that the aggregation is representative of the admin unit as a whole and that the subsequent yield calculation is correct.