Creat AF script
Created an AF script to read the FDW data instead of the AUX data. Note that in this script the data appear as they are originally coded, meaning that there are rainfed and irrigated data instead of data for winter and spring wheat. This means that a comparison to the old verson of the data may be warranted